Rewrite built-ins file names
